+++ twitter4j-core/src/main/java/twitter4j/UserJSONImpl.java
@@ -73,7 +73,7 @@ import static twitter4j.ParseUtil.*;
     private boolean isGeoEnabled;
     private boolean isVerified;
     private boolean isFollowing;
-    private long listedCount;
+    private int listedCount;
     private boolean isFollowRequestSent;
     private static final long serialVersionUID = -6345893237975349030L;
 
@@ -117,7 +117,7 @@ import static twitter4j.ParseUtil.*;
             profileBackgroundTiled = getBoolean("profile_background_tile", json);
             lang = getRawString("lang", json);
             statusesCount = getInt("statuses_count", json);
-            listedCount = getLong("listed_count", json);
+            listedCount = getInt("listed_count", json);
             isFollowRequestSent = getBoolean("follow_request_sent", json);
             if (!json.isNull("status")) {
                 JSONObject statusJSON = json.getJSONObject("status");
@@ -400,7 +400,7 @@ import static twitter4j.ParseUtil.*;
     /**
      * {@inheritDoc}
      */
-    public long getListedCount() {
+    public int getListedCount() {
         return listedCount;
     }
